+++ app/src/main/java/org/breezyweather/sources/WeatherHelper.kt
@@ -101,6 +101,7 @@ class WeatherHelper @Inject constructor(
             }
             mainFeaturesIgnored.add(SecondaryWeatherSourceFeature.FEATURE_ALERT)
         }
+        // TODO: Only call this AFTER main source
         val secondaryWeatherWrapper = requestSecondaryWeather(
             context, location, secondarySources
         )